\name{cluster_pairs}
\alias{cluster_pairs}
\title{Summary of pairs of observations from two different clusterings of the same data set.}
\usage{
  cluster_pairs(cl1, cl2)
}
\arguments{
  \item{cl1}{the cluster labels}

  \item{cl2}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}
\details{
  We assume that the lengths of each vector of cluster
  labels are equal.
}

