\name{clustomit_ramey}
\alias{clustomit_ramey}
\title{Ramey's Alternative Method}
\usage{
  clustomit_ramey(x, K, cluster_method, similarity_method,
    B = 100, with_replacement = TRUE,
    use_multicore = FALSE, ncpus = 1, ...)
}
\arguments{
  \item{x}{data matrix with N observations (rows) and p
  features (columns).}

  \item{num_clusters}{TODO}

  \item{K}{TODO}

  \item{similarity_method}{TODO}

  \item{noise}{TODO}

  \item{B}{TODO}

  \item{use_multicore}{TODO}

  \item{ncpus}{TODO}

  \item{...}{TODO}
}
\value{
  list of scores by omitted cluster
}
\description{
  TODO
}

