\name{hierarchical_wrapper}
\alias{hierarchical_wrapper}
\title{Wrapper for hierarhical clustering, so that only the data set and number of clusters is specified.}
\usage{
  hierarchical_wrapper(x, num_clusters, method = "single",
  ...)
}
\arguments{
  \item{x}{TODO}

  \item{num_clusters}{TODO}

  \item{...}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}
\details{
  The default is single-linkage.
}

