\name{jaccard_indep}
\alias{jaccard_indep}
\title{Computes the Jaccard index for comparing two clusterings of the same data set.}
\usage{
  jaccard_indep(cl1, cl2)
}
\arguments{
  \item{cl1}{TODO}

  \item{cl2}{TODO}
}
\value{
  numeric the Jaccard index for the two sets of cluster
  labels. If an error is encountered, we return NULL.
}
\description{
  TODO
}

