\name{kmeans_wrapper}
\alias{kmeans_wrapper}
\title{Wrapper for k-means, so that only the data set and number of clusters is specified.}
\usage{
  kmeans_wrapper(x, num_clusters, num_starts = 10, ...)
}
\arguments{
  \item{x}{TODO}

  \item{num_clusters}{TODO}

  \item{...}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}

