\name{mclust_wrapper}
\alias{mclust_wrapper}
\title{Wrapper for model-based clustering, so that only the data set and number of clusters is specified.}
\usage{
  mclust_wrapper(x, num_clusters, ...)
}
\arguments{
  \item{x}{TODO}

  \item{num_clusters}{TODO}

  \item{...}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}

