\name{model_wrapper}
\alias{model_wrapper}
\title{Wrapper for model-based clustering, so that only the data set and number of clusters is specified.}
\usage{
  model_wrapper(x, num_clusters, hennig = FALSE, ...)
}
\arguments{
  \item{x}{TODO}

  \item{num_clusters}{TODO}

  \item{hennig}{TODO}

  \item{...}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}
\details{
  To use Hennig's (2007) clusterboot function from the
  'fpc' package with this wrapper for model-based
  clustering, we have to return a custom list to
}

