\name{plot_consensus}
\alias{plot_consensus}
\title{Heatmap of consensus matrix}
\usage{
  plot_consensus(mat)
}
\arguments{
  \item{mat}{the consensus matrix}

  \item{num_clusters}{the number of clusters}

  \item{methods}{a vector of clustering methods to apply in
  consensus}

  \item{...}{(optional) additional arguments to pass to the
  clustering methods}
}
\value{
  TODO
}
\description{
  TODO NOTE: Returns a ggplot2 object. Need to 'print' the
  object.
}

