\name{rand_indep}
\alias{rand_indep}
\title{Computes the Rand index for comparing two clusterings of the same data set.}
\usage{
  rand_indep(cl1, cl2)
}
\arguments{
  \item{cl1}{TODO}

  \item{cl2}{TODO}
}
\value{
  numeric the Rand index for the two sets of cluster
  labels. If an error is encountered, we return NULL.
}
\description{
  TODO
}

