\name{similarity_methods}
\alias{similarity_methods}
\title{Brief description of all similarity functions in the 'clusteval' package.}
\usage{
similarity_methods()
}
\value{
data frame describing each similarity function in the
package. The columns of the data frame are: \describe{
\item{method}{the name of the R function to calculate the
similarity method} \item{name}{The name of the similarity
method} }
}
\description{
For all similarity functions in the \code{\link{clusteval}}
package, we summarize them in a data frame.
}

