\name{singlelinkage_wrapper}
\alias{singlelinkage_wrapper}
\title{Wrapper for single-linkage hierarhical, so that only the data set and number of clusters is specified.}
\usage{
  singlelinkage_wrapper(x, num_clusters, ...)
}
\arguments{
  \item{x}{TODO}

  \item{num_clusters}{TODO}

  \item{...}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}

