% Generated by roxygen2 (4.0.1): do not edit by hand
\name{variation_information}
\alias{variation_information}
\title{Computes the Variation of Information distance between two clusterings}
\usage{
variation_information(labels1, labels2)
}
\arguments{
\item{labels1}{a vector of \code{n} clustering labels}

\item{labels2}{a vector of \code{n} clustering labels}
}
\value{
the VI distance between \code{labels1} and \code{labels2}
}
\description{
This function calculates Meila's (2007) Variation of Information (VI) metric
between two clusterings of the same data set. VI is an information-theoretic
criterion that measures the amount of information lost and gained between two
clusterings.
}
\details{
If \code{n} is the number of observations in the data set, VI is bound
between 0 and \code{log(n)}. Furthermore, VI == 0 if and only if the two
clusterings are the same.

The definition of VI, more properties, and connections to other criteria are
given in the Meila (2007) paper, which has open access:
\url{http://www.sciencedirect.com/science/article/pii/S0047259X06002016}

NOTE: We define 0 log 0 = 0.
}
\examples{
# We generate K = 3 labels for each of n = 30 observations and compute the
# Variation of Information (VI) between the two clusterings.
set.seed(42)
K <- 3
n <- 30
labels1 <- sample.int(K, n, replace=TRUE)
labels2 <- sample.int(K, n, replace=TRUE)
variation_information(labels1, labels2)

# Here, we cluster the \\code{\\link{iris}} data set with the K-means and
# hierarchical algorithms using the true number of clusters, K = 3.
# Then, we compute the VI between the two clusterings.
iris_kmeans <- kmeans(iris[, -5], centers = 3)$cluster
iris_hclust <- cutree(hclust(dist(iris[, -5])), k = 3)
variation_information(iris_kmeans, iris_hclust)
}
\references{
Meila, M. (2007). "Comparing clusterings - an information based
distance," Journal of Multivariate Analysis, 98, 5, 873-895.
\url{http://www.sciencedirect.com/science/article/pii/S0047259X06002016}
}

