\name{vec_equal}
\alias{vec_equal}
\title{Test if all the elements of the vector x are equal.}
\usage{
  vec_equal(x, tol = .Machine$double.eps^0.5)
}
\arguments{
  \item{x}{TODO}

  \item{tol}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}
\details{
  My code is based on the following Stack Overflow post.
  http://stackoverflow.com/questions/4752275/test-for-equality-among-all-elements-of-a-single-vector
}

