% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clusteval.r
\docType{package}
\name{clusteval}
\alias{clusteval}
\alias{clusteval-package}
\alias{package-clusteval}
\title{Evaluation of Clustering Algorithms}
\description{
An R package that provides a suite of tools to evaluate clustering algorithms,
clusterings, and individual clusters.
}

