\name{comembership}
\alias{comembership}
\title{Co-Membership}
\usage{
  comembership(clust_labels)
}
\arguments{
  \item{clust_labels}{a vector of clustering labels}
}
\value{
  a vector of co-membership bits
}
\description{
  For a set of clustering labels, computes the
  co-membership of all pairs of observations. Basically,
  the co-membership is defined as the pairs of observations
  that are clustered together.
}
\details{
  TODO: Add purpose. (Label switching problem). TODO: Add
  Kapp and Tibshirani (2007) Biostatistics paper as
  reference on 'co-membership' term. TODO: Add to the name
  of the function.
}
\examples{
TODO
}

