\name{diana_wrapper}
\alias{diana_wrapper}
\title{Wrapper for DIANA, so that only the data set and number of clusters is specified.}
\usage{
  diana_wrapper(x, num_clusters, hennig = FALSE, ...)
}
\arguments{
  \item{x}{TODO}

  \item{num_clusters}{TODO}

  \item{...}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}

